requirements.txt
aiogram
aiohttp
websockets
pydantic
python-dotenv
pyyaml
redis
ujson
